Move plugins to their own folder
