+++ src/plugins/coincap.js
